Add files via upload
I wanted to give you a working copy of the original template as it has many of the features that you are currently working on and inventing, you just need a working copy, but the files were too big so I split the app and the database and so when you unzip these files just put the database files back into the data folder not the database folder and all should be well. speak soon.